pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-assistant-rag"
version = "0.1.0"
description = "Generic email understanding and routing assistant using LangChain, Mistral AI, and FAISS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-mistralai>=0.2.0",
  "faiss-cpu>=1.8.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "black>=24.4.0", "ruff>=0.5.0"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
